Convert to AdoptOpenJDK
This changes the base image to the one from AdoptOpenJDK's latest JDK (they don't provide a tag for 'JRE-latest'). AdoptOpenJDK is currently at 15.x, which includes Shenandoah GC and other improvements, and is based on Ubuntu images just like the 'multiarch' image is.

I'll issue some PRs for other changes that actually make use of the new features shortly.