Envvars in files from `/config` and `/mods` now replaced as expected
The operation that copies files from `/config` and `/mods` mounts were
previously executed in the final step, so env vars copied from these
mounts did not get replaced (#700). Moved these operations to the file
which contains the operation that copies `/plugins`. This file was
renamed to "start-finalSetupMounts", to get all these together.