Add support for Canyon server for b1.7.3
Canyon is a fork of CraftBukkit for Minecraft Beta 1.7.3, created by
some of the Velocity team members. It includes multiple enhancements
whilst also retaining compatibility with old Bukkit plugins and mods as
much as possible.

This commit adds initial support for this project, allowing for
downloading of the latest successful build, similar to the Airplane
support method. Future commits will add support for specific build or
version retrieval.