Determine Canyon build version and download as appropriate
This commit adds support for determining the build version of Canyon
downloaded from Velocity's CI server.

As a method of version detection, build tags are added to the untagged
jar (which is the current behavior) so that downloading of newer
versions when the latest successful build is requested occurs
automatically. This also removes the use of FORCE_REDOWNLOAD as it is
unnecessary with this functon.