document that default RCON_PASSWORD = minecraft in README
because while experimenting with RCON, it's highly confusing why without a RCON_PASSWORD it happens
to "just work" e.g. with https://github.com/willroberts/minecraft-client (which uses "minecraft" as the default password itself!)
but not with https://github.com/itzg/rcon-cli (which has no default password).

Default passwords are always a bad idea, and should ideally ideally just be removed (if you are willing to make that backward compatible breaking change).