Implement support for ops/whitelist json
This commit adds support for providing the JSON file used for ops and
whitelist definitions. This is provided via the OPS_JSON and
WHITELIST_JSON environment variables, and similar to OPS and WHITELIST
they are copied on first run and only override the existing files if
OVERRIDE_OPS and OVERRIDE_WHITELIST are set to true.

If both forms of definition are provided, the JSON is processed first,
with additional processing occuring with the user/uuid CSV variables.
This ensures consistent behaviour with hiw options are currently used,
and allows for override configuration to be provided at runtime or
definition (for example a default whitelist/ops configuration, with
additionals added for an event)