Handle pre-json whitelist/ops configuration better
Minecraft 1.7.6 added support for UUIDs via the JSON configuration file
format for whitelist and ops. To handle 1.7.5 and earlier, we perform
slightly different logic.

If Minecraft 1.7.5 or earlier:
- Delete the correct files for OVERRIDE_OPS/OVERRIDE_WHITELIST
- Update ops.txt and white-list.txt with users provided via variables
- Error on use of JSON environtment variables

As part of this change, a fix to handle existing configuration for both
JSON and text-based configurations was added. If existing files are
found, we uniquely merge with them.