Rewrite parser for OPS and WHITELIST
The parser has been rewritten to individually work on usernames/UUIDs
passed in the OPS and WHITELIST environment variables.

If a username or UUID could not be looked up, we log a warning about the
error. No failure state is created and the user is skipped.

We no longer use the legacy .txt conversion for Minecraft >= 1.7.6, we
now update the ops.json or whitelist.json automatically.

For both old txt and new json formats we ensure uniqueness.

As part of this change, support for UUIDs in OPS has been added.