Clarify use of whitelist and enforcement
Documentation has been updated to draw attention to the requirement of
ENFORCE_WHITELIST to actually enforce the configured whitelist. A server
log message has been added to warn if the whitelist functionality is
enabled but not enforced.

The duplicate setting of 'whitelist' and 'white-list' in
server.properties has been removed, leaving only 'white-list' being
applied. I believe this was an error introduced with the addition of
whitelist functionality in 111883e, that was subsequently fixed in
32cb5f4.