Adjust special case for Spigot world layout
Spigot-family servers put the nether and end dimensions into separate
world directories. Previously, the start-setupWorld script only had
support for this layout when loading a Spigot-family jar. This commit
adjusts the approach for detecting the Spigot layout, so that even when
loading a vanilla jar, the correct world can be identified.

Note that this commit does *not* fix actually loading Spigot nether and
end dimensions on vanilla jars; it merely improves the detection for the
main world directory.