Migrate Fabric server to launcher method
The Fabric team have improved the server launch methodology to be more
in line with other server types - a launcher specifying the Minecraft,
Fabric Launcher, and Fabric Loader versions to use.

https://fabricmc.net/use/server/

This allows for cleaner auto-updates, and fixes #1350 by using the
naming convention Fabric use for launcher jar files.

Documentation has been updated to reflect this change, and support for
existing configurations is maintained, with the exception of
FABRIC_INSTALLER and FABRIC_INSTALLER_URL (as the installer is no longer
used).