Download Canyon from GitHub for final build
Canyon is on an indefinite hiatus, and a final build has been produced
by Kobo for the meantime. This changes the default value for
CANYON_BUILD to "final", and in the instance this is provided, we
download this build from GitHub rather than the CI integration.

The CI server previously provided by Velocity is now 410 Gone, so
individual build numbers are not available at present. A future PR may
address this by them being made available by another provider, but for
now I think this change is enough to keep Canyon servers functional for
the meantime.